refactor(terminal): extract per-session hook from terminal-dock (no behaviour change)
Extract src/components/board/use-terminal-session.ts out of terminal-dock.tsx
(1508 -> 631 lines): a hook that owns everything ONE session needs — the
terminalReducer instance, WebSocket browser leg, xterm + fit addon, heartbeat +
silent-link watchdog, grace-window reattach loop, read-only gating, resize and
focus handling, and the vibecodes:// deep-link fire. terminal-dock.tsx becomes a
consumer for its single session, keeping the dock CHROME (expanded/collapsed,
launch-bus wiring, all rendering) and calling `requestExpand()` at the points it
used to call `setExpanded(true)` directly. src/lib/terminal/connection.ts is
untouched; its existing 51 tests pass unmodified.

Q1 verdict (helper/main.js, report-only per stage scope): concurrent
vibecodes:// launches are ALREADY handled correctly — no fix needed. Electron's
requestSingleInstanceLock (main.js:233-240) plus the open-url/second-instance
handlers (line 246, 237-240) both funnel into handleLaunchUrl, which `fork`s a
NEW independent bridge child per call (line 175-186) into a `children` Set with
no serialization/mutex; the idle-quit (94-108) only fires when the Set AND
activeReaps are both empty, so N concurrent launches -> N concurrent bridges
under one helper process, each tracked independently via ptyPids keyed by child.

New unit tests (use-terminal-session.test.ts, 7 cases) cover the hook's wiring:
mint -> open -> connected, read-only gating, user-end, mint failure, and the
grace-window reattach (including reconnect-exhausted -> honest session-ended).

Gates: npx tsc --noEmit clean; npx vitest run src/lib/terminal/ src/components/board/
-> 22 files / 217 tests pass; npm run lint -> 0 errors (77 pre-existing warnings,
none new, none in the touched files); terminal/test suite -> 72/72 pass.
